Fix all ruff lint findings across app/ and tests/
Cleared ~280 findings: dead while/else, redundant int() cast, logger-level
timestamp not tz-aware, .error(exc_info=True) -> .exception(), two silently
swallowed websocket recv failures given debug logging for parity with their
sibling send-path logging, nested ifs/withs merged, unused loop variable,
list()[0] -> next(iter()), generator -> set comprehension, dict .values()
over .items(), unused tuple-unpack bindings prefixed with _, import sorting,
and justified `# noqa` on best-effort I/O/network/cleanup catches that are
already logged or are deliberate fuzz-test "any non-crash is fine" catches.

No functional changes; full suite (3326 tests, 19 subtests) still green.

+++ app/nibe_entity_detection.py
@@ -391,7 +391,7 @@ def reverse_divisor(display_value: float, divisor: int) -> int:
     the full rationale.
     """
     effective = divisor if divisor and divisor != 0 else 1
-    return int(round(display_value * effective))
+    return round(display_value * effective)
 
 
 def create_entity_id(point_id: int) -> str:
